Adding esm module to use both import & require
