+++ package.json
@@ -67,6 +67,7 @@
     "date-holidays": "^3.9.1",
     "electron-squirrel-startup": "^1.0.0",
     "electron-store": "^6.0.1",
+    "esm": "^3.2.25",
     "i18next": "^21.2.3",
     "i18next-node-fs-backend": "^2.1.3",
     "is-online": "^9.0.1",
